+++ cmd/internal/cli/actions.go
@@ -103,6 +103,7 @@ func handleOCI(ctx context.Context, imgCache *cache.Handle, cmd *cobra.Command, 
 		DockerHost:  dockerHost,
 		NoHTTPS:     noHTTPS,
 		OciSif:      isOCI,
+		KeepLayers:  keepLayers,
 		ReqAuthFile: reqAuthFile,
 	}
 
@@ -143,6 +144,7 @@ func handleLibrary(ctx context.Context, imgCache *cache.Handle, pullFrom string)
 		LibraryConfig: c,
 		// false to allow OCI execution of native SIF from library
 		RequireOciSif: false,
+		KeepLayers:    keepLayers,
 		TmpDir:        tmpDir,
 		Platform:      getOCIPlatform(),
 	}